+++ src/flowfield/spectral_solver.rs
@@ -17,27 +17,27 @@ use std::sync::Arc;
 use crate::{BoxSize, GridSize};
 
 pub struct SpectralSolver {
-    flow_field: Array<Complex<f64>, Ix4>,
+    flow_field: Array<Complex<f32>, Ix4>,
     fft_plan_forward: Arc<FFTPlan>,
     fft_plan_backward: Arc<FFTPlan>,
-    k_invnormsquared: Array<Complex<f64>, Ix3>,
-    k_mesh: Array<Complex<f64>, Ix4>,
-    k_normed_mesh: Array<Complex<f64>, Ix4>,
-    stress_kernel: Array<f64, Ix4>,
-    stress_field: Array<Complex<f64>, Ix5>,
-    gradient_meanf: Array<Complex<f64>, Ix5>,
+    k_invnormsquared: Array<Complex<f32>, Ix3>,
+    k_mesh: Array<Complex<f32>, Ix4>,
+    k_normed_mesh: Array<Complex<f32>, Ix4>,
+    stress_kernel: Array<f32, Ix4>,
+    stress_field: Array<Complex<f32>, Ix5>,
+    gradient_meanf: Array<Complex<f32>, Ix5>,
 }
 
 impl SpectralSolver {
     pub fn new<F>(grid_size: GridSize, box_size: BoxSize, stress: F) -> SpectralSolver
     where
-        F: Fn(f64, f64) -> Array<f64, Ix2>,
+        F: Fn(f32, f32) -> Array<f32, Ix2>,
     {
         let grid_width = GridWidth::new(grid_size, box_size);
 
         let mesh = get_k_mesh(grid_size, box_size);
 
-        let mut dummy: Array<Complex<f64>, Ix3> =
+        let mut dummy: Array<Complex<f32>, Ix3> =
             Array::default([grid_size.x, grid_size.y, grid_size.z]);
         let plan_stress = FFTPlan::new_c2c_inplace_3d(
             &mut dummy.view_mut(),
@@ -120,13 +120,13 @@ impl SpectralSolver {
             .into_par_iter()
             .for_each(|mut v| fft.reexecute3d(&mut v));
 
-        let norm = (dist_sh.0 * dist_sh.1 * dist_sh.2) as f64;
+        let norm = (dist_sh.0 * dist_sh.1 * dist_sh.2) as f32;
 
         // convert to real vector field
         u.map(|v| v.re / norm)
     }
 
-    pub fn fft_mean_flow_field(&mut self, screening: f64, dist: &Distribution) {
+    pub fn fft_mean_flow_field(&mut self, screening: f32, dist: &Distribution) {
         let dist_sh = dist.dim();
         let stress_sh = self.stress_kernel.dim();
         let n_stress = stress_sh.0 * stress_sh.1;
@@ -162,7 +162,7 @@ impl SpectralSolver {
         let ik2 = self.k_invnormsquared.view();
         let ik2 = ik2.into_shape([n]).unwrap();
 
-        let norm = n as f64;
+        let norm = n as f32;
 
         Zip::from(ff.axis_iter_mut(Axis(1)))
             .and(stress_field.axis_iter(Axis(2)))
@@ -228,9 +228,9 @@ impl SpectralSolver {
     /// field and the (flattened) vector gradient field of it.
     pub fn mean_flow_field(
         &mut self,
-        screening: f64,
+        screening: f32,
         d: &Distribution,
-    ) -> (ArrayView<Complex<f64>, Ix4>, ArrayView<Complex<f64>, Ix5>) {
+    ) -> (ArrayView<Complex<f32>, Ix4>, ArrayView<Complex<f32>, Ix5>) {
         // calculate FFT of of flow field, which is stored in self.flow_field
         self.fft_mean_flow_field(screening, d);
         // use stored value of FFT of magnetic field to calculate vector gradient and
@@ -246,7 +246,7 @@ impl SpectralSolver {
         (self.flow_field.view(), self.gradient_meanf.view())
     }
 
-    pub fn get_real_flow_field(&self) -> Array<f64, Ix4> {
+    pub fn get_real_flow_field(&self) -> Array<f32, Ix4> {
         self.flow_field.map(|v| v.re)
     }
 }
