+++ src/flowfield/stress.rs
@@ -8,24 +8,24 @@ use crate::distribution::Distribution;
 use crate::mesh::grid_width::GridWidth;
 use ndarray::{Array, ArrayView, ArrayViewMut, Axis, Ix2, Ix4, Ix5};
 use num_complex::Complex;
-use std::f64::consts::PI;
+use std::f32::consts::PI;
 use crate::GridSize;
 
 /// Holds prefactors for active and magnetic stress
 #[derive(Debug, Copy, Clone, Serialize, Deserialize)]
 #[serde(deny_unknown_fields)]
 pub struct StressPrefactors {
-    pub active: f64,
-    pub magnetic: f64,
+    pub active: f32,
+    pub magnetic: f32,
 }
 
 /// Calculates approximation of discretized stress kernel, to be used in ///
 /// the expectation value to obtain the stress tensor.
-pub fn stress_kernel<F>(grid_size: GridSize, grid_width: GridWidth, stress: F) -> Array<f64, Ix4>
+pub fn stress_kernel<F>(grid_size: GridSize, grid_width: GridWidth, stress: F) -> Array<f32, Ix4>
 where
-    F: Fn(f64, f64) -> Array<f64, Ix2>,
+    F: Fn(f32, f32) -> Array<f32, Ix2>,
 {
-    let mut s = Array::<f64, _>::zeros((3, 3, grid_size.phi, grid_size.theta));
+    let mut s = Array::<f32, _>::zeros((3, 3, grid_size.phi, grid_size.theta));
     // Calculate discrete angles, considering the cell centered sample points of
     // the distribution
     let gw_half_phi = grid_width.phi / 2.;
@@ -53,10 +53,10 @@ where
 /// It consumes `stress_field` and updates it given a stress kernel `kernel`
 /// and a distribution `dist`. It returns the updated stress field.
 pub fn average_stress<'a>(
-    stress_field: ArrayViewMut<'a, Complex<f64>, Ix5>,
-    kernel: &ArrayView<f64, Ix4>,
+    stress_field: ArrayViewMut<'a, Complex<f32>, Ix5>,
+    kernel: &ArrayView<f32, Ix4>,
     dist: &Distribution,
-) -> ArrayViewMut<'a, Complex<f64>, Ix5> {
+) -> ArrayViewMut<'a, Complex<f32>, Ix5> {
     let dist_sh = dist.dim();
     let stress_sh = kernel.dim();
 
@@ -94,7 +94,7 @@ pub mod stresses {
     use ndarray::{Array, Ix2};
 
     /// Calculate active stress tensor for polar angles `phi` and `theta`.
-    pub fn stress_active(phi: f64, theta: f64) -> Array<f64, Ix2> {
+    pub fn stress_active(phi: f32, theta: f32) -> Array<f32, Ix2> {
         let mut s = Array::zeros((3, 3));
 
         s[[0, 0]] = -(1. / 3.) + phi.cos() * phi.cos() * theta.sin() * theta.sin();
@@ -116,7 +116,7 @@ pub mod stresses {
     /// Magnetic field points into y-direction to avoid binning instability in
     /// spherical coordinates at the poles.
     /// Calculates `0.5 (nb-bn)`, with `b = [0, 1, 0]` and orientation `n`.
-    pub fn stress_magnetic(phi: f64, theta: f64) -> Array<f64, Ix2> {
+    pub fn stress_magnetic(phi: f32, theta: f32) -> Array<f32, Ix2> {
         let mut s = Array::zeros((3, 3));
 
         s[[0, 0]] = 0.;
@@ -135,7 +135,7 @@ pub mod stresses {
     }
 
     /// Calculate magnetic stress tensor for polar angles `phi` and `theta`.
-    pub fn stress_magnetic_rods(phi: f64, theta: f64) -> Array<f64, Ix2> {
+    pub fn stress_magnetic_rods(phi: f32, theta: f32) -> Array<f32, Ix2> {
         let mut s = Array::zeros((3, 3));
 
         s[[0, 0]] = 2. * theta.cos() * phi.cos().powi(2) * theta.sin().powi(2);
